fix: fixes provided config being undefined
