+++ package.json
@@ -15,10 +15,6 @@
       "types": "./dist/types.d.ts",
       "import": "./dist/module.mjs",
       "require": "./dist/module.cjs"
-    },
-    "./config": {
-      "import": "./dist/index.mjs",
-      "require": "./dist/index.cjs"
     }
   },
   "main": "./dist/module.cjs",
