+++ playground/sanctum.config.ts
@@ -1,3 +1,4 @@
 export default {
     baseUrl: 'http://localhost:80',
+    userStateKey: 'sanctum.user.identity',
 };
