--- src/index.ts
@@ -1,5 +1,0 @@
-import type { SanctumConfigOptions } from './types';
-
-export const defineSanctumConfig = (config: Partial<SanctumConfigOptions>) => config;
-
-export type { SanctumConfigOptions };
