+++ src/module.ts
@@ -1,3 +1,4 @@
+import { existsSync } from 'node:fs';
 import type { SanctumModuleOptions } from './types';
 import {
     defineNuxtModule,
@@ -5,8 +6,8 @@ import {
     createResolver,
     addImportsDir,
     addRouteMiddleware,
+    addPluginTemplate,
 } from '@nuxt/kit';
-import { addNuxtAuthSanctumConfig } from './config';
 
 export default defineNuxtModule<SanctumModuleOptions>({
     meta: {
@@ -21,10 +22,56 @@ export default defineNuxtModule<SanctumModuleOptions>({
         configFile: 'sanctum.config.ts',
     },
 
-    setup(options) {
+    setup(options, nuxt) {
         const resolver = createResolver(import.meta.url);
 
-        addNuxtAuthSanctumConfig(options);
+        const configBase = resolver.resolve(nuxt.options.rootDir, options.configFile ?? 'sanctum.config');
+
+        addPlugin(resolver.resolve('./runtime/plugin'));
+
+        addPluginTemplate({
+            filename: 'sanctum-config.mjs',
+            async getContents() {
+                const configPath = await resolver.resolvePath(configBase);
+                const configPathExists = existsSync(configBase);
+
+                return `
+                    import { defineNuxtPlugin } from '#imports';
+                    ${configPathExists ? "import defu from 'defu';" : ''}
+                    ${configPathExists ? `import sanctumConfig from '${configPath}'` : ''}
+
+                    export default defineNuxtPlugin((nuxtApp) => {
+                        const defaultConfig = {
+                            userStateKey: 'sanctum.user.identity',
+                            redirectIfAuthenticated: false,
+                            endpoints: {
+                                csrf: '/sanctum/csrf-cookie',
+                                login: '/login',
+                                logout: '/logout',
+                                user: '/api/user',
+                            },
+                            csrf: {
+                                cookie: 'XSRF-TOKEN',
+                                header: 'X-XSRF-TOKEN',
+                            },
+                            client: {
+                                retry: false,
+                            },
+                            redirect: {
+                                keepRequestedRoute: false,
+                                onLogin: '/',
+                                onLogout: '/',
+                                onAuthOnly: '/login',
+                                onGuestOnly: '/',
+                            },
+                        };
+
+                        const config = ${configPathExists ? `defu(sanctumConfig, defaultConfig)` : `defaultConfig`};
+                        nuxtApp.provide('sanctumConfig', config);
+                    });
+                `;
+            },
+        });
 
         addImportsDir(resolver.resolve('./runtime/composables'));
 
@@ -36,7 +83,5 @@ export default defineNuxtModule<SanctumModuleOptions>({
             name: 'sanctum:guest',
             path: resolver.resolve('./runtime/middleware/sanctum.guest'),
         });
-
-        addPlugin(resolver.resolve('./runtime/plugin'));
     },
 });
